# 🚧 Optimise stylesheet and JavaScript assets