--- a/app/assets/javascripts/application.js
+++ b/app/assets/javascripts/application.js
@@ -1,2 +1,2 @@
-//= require govuk_publishing_components/dependencies
-//= require govuk_publishing_components/all_components
+// = require govuk/all.js
+window.GOVUKFrontend.initAll()