Selectively import stylesheets and only import and initialise specific script.

In the table below a comparison of sizes (uncompressed) and localhost times.
<table>
<tr><th></th><th>Size Before</th><th> Size After</th><th>Time Before</th><th> Time After</th></tr>
<tr><th align="left">application.js</th>
  <td>644 KB</td><td>80.6 KB</td>
  <td>351 ms</td><td>19 ms</td>
</tr>
<tr><th align="left">application.css</th>
  <td>825 KB</td><td>317 KB</td>
  <td>356 ms</td><td>35 ms</td>
</tr>
</table>

### Before
<img width="1440" alt="assets-before" src="https://user-images.githubusercontent.com/788096/77944543-0f548380-72b7-11ea-8c4d-d41838ddb918.png">


### After
<img width="1440" alt="assets-after" src="https://user-images.githubusercontent.com/788096/77944559-124f7400-72b7-11ea-8a6e-05ea6f51e2e9.png">

This also removes jQuery, which means an overall audit should show significant improvement – will update with a Lighthouse report in production after it gets merged.

[Trello card](https://trello.com/c/I35qjJcP)